move Dictionary into rotors package
