+++ src/rotors/Rotor.java
@@ -23,7 +23,6 @@ package rotors;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import enigma.Dictionary;
 import interfaces.RotaryEncryptor;
 
 public class Rotor implements RotaryEncryptor {
